+++ app/src/main/java/com/paula/android/bechef/data/entity/BookmarkItem.java
@@ -4,7 +4,7 @@ import androidx.room.Entity;
 
 @Entity(tableName = "bookmark_item")
 public class BookmarkItem extends BaseItem {
-    public BookmarkItem(int tabUid, String title, String imageUrl, double rating,
+    public BookmarkItem(long tabUid, String title, String imageUrl, double rating,
                         String tags, int inTodayId, String createdTime, String description, String videoId) {
         super(tabUid, title, imageUrl, rating, tags, inTodayId, createdTime, description, videoId);
     }
